city-guide: move utils out of pages
previously, the util in pages was causing the deploy to break
