+++ city-guide/utils/images.ts
